fixed audio added enemies

+++ GameEngine2/TileMap.cpp
@@ -24,8 +24,8 @@ namespace nc {
 			if (index != 0) {
 				GameObject* gameObject = ObjectFactory::Instance().Create<GameObject>(m_tileNames[index]);
 
-				float x = i % m_numX;
-				float y = i / m_numX;
+				float x = static_cast<float>(i % m_numX);
+				float y = static_cast<float>(i / m_numX);
 				gameObject->m_transform.position = m_start + nc::Vector2{ x, y } * m_size;
 
 				scene->AddGameObject(gameObject);
